🏗️ refactor(core): move core files from features to core directory
Relocates auth context and HTTP client to core directory for better architecture organization and import paths alignment.

Changes include:
- Move AuthContext to core/contexts
- Move AvocadoClient to core/https
- Update all related imports
- Remove unused mock code and interfaces
