+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { AuthProvider } from "@/features/auth/contexts/AuthContext";
+import { AuthProvider } from "@/core/contexts/AuthContext";
 
 const inter = Inter({ subsets: ["latin"] });
