+++ src/features/auth/api/login.ts
@@ -15,20 +15,6 @@ import { LoginResponse } from "../types";
  * and return mock response for UI/UX development
  */
 export const login = async (username: string, password: string): Promise<LoginResponse> => {
-  // Check if we should bypass auth in development
-  if (process.env.NEXT_PUBLIC_BYPASS_AUTH === 'true') {
-    console.log('Development mode: Bypassing authentication');
-    return {
-      success: true,
-      content: {
-        access_token: "mock_token_for_development",
-        token_type: "bearer"
-      },
-      message: "Login successfully"
-    };
-  }
-
-  // Normal authentication flow
   const baseURL = process.env.NEXT_PUBLIC_API_BASE_URL || '';
   const login_api = `${baseURL}/api/auth/login`;
