+++ src/features/manage_center/managecenter/updateLicense.tsx
@@ -1,4 +1,4 @@
-import { avocadoClient } from '@/features/api/AvocadoClient';
+import { avocadoClient } from '@/core/https/AvocadoClient';
 
 type LicenseDataType = {
   user_id: number;
